Fixin phpstan errors (task #8244)

+++ src/Controller/SurveysController.php
@@ -11,6 +11,7 @@
  */
 namespace Qobo\Survey\Controller;
 
+use Cake\Datasource\EntityInterface;
 use Cake\Event\Event;
 use Cake\ORM\TableRegistry;
 use Qobo\Survey\Controller\AppController;
@@ -22,6 +23,7 @@ use Qobo\Survey\Event\EventName;
  * @property \Qobo\Survey\Model\Table\SurveysTable $Surveys
  * @property \Qobo\Survey\Model\Table\SurveyAnswersTable $SurveyAnswers
  * @property \Qobo\Survey\Model\Table\SurveyQuestionsTable $SurveyQuestions
+ * @property \Qobo\Survey\Model\Table\SurveyResultsTable $SurveyResults
  *
  * @method \Qobo\Survey\Model\Entity\Survey[]|\Cake\Datasource\ResultSetInterface paginate($object = null, array $settings = [])
  */
@@ -89,7 +91,9 @@ class SurveysController extends AppController
         if (!empty($event->result)) {
             $submits = $event->result;
         } else {
-            $submits = $this->SurveyResults->getSubmits($survey->id);
+            if (! empty($survey)) {
+                $submits = $this->SurveyResults->getSubmits($survey->get('id'));
+            }
         }
 
         $this->set(compact('survey', 'questionTypes', 'submits'));
@@ -98,12 +102,18 @@ class SurveysController extends AppController
     /**
      * Publish method
      *
-     * @param string|null $id Survey id.
+     * @param string $id Survey id.
      * @return \Cake\Http\Response|void|null Redirects on successful add, renders view otherwise.
      */
-    public function publish(string $id = null)
+    public function publish(string $id)
     {
         $survey = $this->Surveys->getSurveyData($id);
+
+        if (empty($survey)) {
+            $this->Flash->error((string)__('Couldn\'t find a survey with given id'));
+
+            return $this->redirect($this->referer());
+        }
 
         if ($this->request->is(['post', 'put', 'patch'])) {
             $validated = $this->Surveys->prepublishValidate($id, $this->request);
@@ -115,7 +125,7 @@ class SurveysController extends AppController
             }
 
             $data = $this->request->getData();
-            $survey = $this->Surveys->patchEntity($survey, $data, ['validate' => false]);
+            $survey = $this->Surveys->patchEntity($survey, (array)$data, ['validate' => false]);
 
             if ($this->Surveys->save($survey)) {
                 $this->Flash->success((string)__('Survey was successfully saved.'));
@@ -151,7 +161,12 @@ class SurveysController extends AppController
         $survey = $this->Surveys->getSurveyData($id, true);
 
         if ($this->request->is(['post', 'put', 'patch'])) {
-            $this->SurveyResults = TableRegistry::get('Qobo/Survey.SurveyResults');
+            /**
+             * @var \Qobo\Survey\Model\Table\SurveyResultsTable $table
+             */
+            $table = TableRegistry::getTableLocator()->get('Qobo/Survey.SurveyResults');
+            $this->SurveyResults = $table;
+
             $user = $this->Auth->user();
             $requestData = $this->request->getData();
             if (empty($requestData['SurveyResults'])) {
@@ -202,20 +217,30 @@ class SurveysController extends AppController
         $this->autoRender = false;
         $survey = $this->Surveys->getSurveyData($id);
 
+        if (empty($survey)) {
+            $this->Flash->error((string)__('Couldn\'t duplicate the survey data'));
+
+            return $this->redirect($this->referer());
+        }
+
         if ($this->request->is(['post', 'put', 'patch'])) {
-            $duplicated = $this->Surveys->duplicate($survey->id);
+            /**
+             * @var \Cake\Datasource\EntityInterface
+             */
+            $duplicated = $this->Surveys->duplicate($survey->get('id'));
 
-            if ($duplicated) {
+            if (! empty($duplicated)) {
                 // @NOTE: saving parent_id as Duplicatable unsets origin id
-                $duplicated = $this->Surveys->patchEntity($duplicated, ['parent_id' => $survey->id]);
+                $duplicated = $this->Surveys->patchEntity($duplicated, ['parent_id' => $survey->get('id')]);
                 $duplicated = $this->Surveys->save($duplicated);
+                if ($duplicated instanceof EntityInterface) {
+                    // Fixing order from original survey if there were any order gaps.
+                    $sorted = $this->Surveys->setSequentialOrder($duplicated);
 
-                // Fixing order from original survey if there were any order gaps.
-                $sorted = $this->Surveys->setSequentialOrder($duplicated);
+                    $this->Flash->success((string)__('Survey was successfully duplicated'));
 
-                $this->Flash->success((string)__('Survey was successfully duplicated'));
-
-                return $this->redirect(['action' => 'view', $duplicated->id]);
+                    return $this->redirect(['action' => 'view', $duplicated->get('id')]);
+                }
             }
             $this->Flash->error((string)__('Couldn\'t duplicate the survey data'));
         }
@@ -230,7 +255,7 @@ class SurveysController extends AppController
         $survey = $this->Surveys->newEntity();
 
         if ($this->request->is('post')) {
-            $survey = $this->Surveys->patchEntity($survey, $this->request->getData());
+            $survey = $this->Surveys->patchEntity($survey, (array)$this->request->getData());
             if ($this->Surveys->save($survey)) {
                 $this->Flash->success((string)__('The survey has been saved.'));
 
@@ -244,13 +269,19 @@ class SurveysController extends AppController
     /**
      * Edit method
      *
-     * @param string|null $id Survey id.
+     * @param string $id Survey id.
      * @return \Cake\Http\Response|void|null Redirects on successful edit, renders view otherwise.
      */
-    public function edit(string $id = null)
+    public function edit(string $id)
     {
         $survey = $this->Surveys->getSurveyData($id);
-        $redirect = ['controller' => 'Surveys', 'action' => 'view', $survey->id];
+        $redirect = ['controller' => 'Surveys', 'action' => 'view', $id];
+
+        if (empty($survey)) {
+            $this->Flash->error((string)__('Couldn\'t find Survey with given id'));
+
+            return $this->redirect($this->referer());
+        }
 
         if (!empty($survey->publish_date)) {
             $this->Flash->error((string)__('You cannot edit alredy published survey'));
@@ -259,7 +290,7 @@ class SurveysController extends AppController
         }
 
         if ($this->request->is(['patch', 'post', 'put'])) {
-            $survey = $this->Surveys->patchEntity($survey, $this->request->getData());
+            $survey = $this->Surveys->patchEntity($survey, (array)$this->request->getData());
             if ($this->Surveys->save($survey)) {
                 $this->Flash->success((string)__('The survey has been saved.'));
 
@@ -280,6 +311,13 @@ class SurveysController extends AppController
     {
         $this->request->allowMethod(['post', 'delete']);
         $survey = $this->Surveys->getSurveyData($id);
+
+        if (empty($survey)) {
+            $this->Flash->error((string)__('The survey could not be deleted. Please, try again.'));
+
+            return $this->redirect(['action' => 'index']);
+        }
+
         if ($this->Surveys->delete($survey)) {
             $this->Flash->success((string)__('The survey has been deleted.'));
         } else {
@@ -299,18 +337,32 @@ class SurveysController extends AppController
      */
     public function viewSubmit(string $surveyId, string $submitId)
     {
-        $this->SurveyQuestions = TableRegistry::get('Qobo/Survey.SurveyQuestions');
+        $surveyResults = [];
+        $survey = null;
+
+        /**
+         * @var \Qobo\Survey\Model\Table\SurveyQuestionsTable $table
+         */
+        $table = TableRegistry::get('Qobo/Survey.SurveyQuestions');
+        $this->SurveyQuestions = $table;
+
         $survey = $this->Surveys->getSurveyData($surveyId);
 
-        $query = $this->SurveyQuestions->find()
-            ->where(['survey_id' => $survey->id])
-            ->contain([
-                'SurveyAnswers' => ['SurveyResults' => function ($q) use ($submitId) {
-                    return $q->where(['submit_id' => $submitId]);
-                }]
-            ]);
+        if (! empty($survey)) {
+            $query = $this->SurveyQuestions->find()
+                ->where(['survey_id' => $survey->get('id')]);
+            $query->enableHydration(true);
+            $query->contain([
+                    'SurveyAnswers' => ['SurveyResults' => function ($q) use ($submitId) {
+                        return $q->where(['submit_id' => $submitId]);
+                    }]
+                ]);
 
-        $surveyResults = $query->all();
+            if ($query->count()) {
+                $surveyResults = $query->all();
+            }
+        }
+
         $this->set(compact('survey', 'surveyResults'));
     }
 }
